requirements.txt
python-dotenv
groq
jinja2
scikit-learn
numpy
